minor adjustments
The order of commands was a little bit wrong, so I changed it. Also as of now I will split up the instructions into different ones. One for example for ext4 and one for btrfs.

I will provide instructions for rEFInd aswell.

Also ricing and vfio will be completed.

I am using my new Arch install for a month now and gained lots of experience by doing so.